chore: stubbing out initial chat with hardcoded user
