+++ react-native/ModeratedChat/App.tsx
@@ -1,5 +1,5 @@
+// import MomentoTestApp from './MomentoTestApp';
 import ChatApp from './ChatApp';
-
 
 export default function App() {
   return (<ChatApp/>);
